backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.7.0
pydantic-settings>=2.2.1
httpx>=0.27.0
python-dotenv>=1.0.1
gitpython>=3.1.43
docker>=7.1.0
kubernetes>=30.1.0
jinja2>=3.1.4
firebase-admin>=6.5.0
pyyaml>=6.0.1
pinecone>=3.0.0
requests>=2.31.0
groq>=0.4.0
requests
